requirements.txt
absl-py==0.15.0
astunparse==1.6.3
beartype==0.10.4
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.12
colored==1.4.4
commonmark==0.9.1
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
flatbuffers==1.12
fuzzywuzzy==0.18.0
gast==0.3.3
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
humanfriendly==10.0
humanize==3.14.0
idna==3.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.1
Keras==2.4.3
keras-bert==0.86.0
keras-embed-sim==0.10.0
keras-layer-normalization==0.16.0
keras-multi-head==0.29.0
keras-pos-embd==0.13.0
keras-position-wise-feed-forward==0.8.0
Keras-Preprocessing==1.1.2
keras-self-attention==0.51.0
keras-transformer==0.40.0
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.2.2
opencv-python==4.3.0.38
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.14.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
requests==2.27.1
requests-oauthlib==2.0.0
rich==12.0.1
rsa==4.9.1
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.15.0
slurm-gpustat==0.0.15
tabulate==0.8.10
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.14.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
threadpoolctl==3.1.0
tqdm==4.64.1
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.20
validators==0.20.0
vit-keras==0.1.0
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
